Preparatory work for calling Instrument.Argument.
Add some stuff to Signature to deal with LocalVarSig.
This is because we're going to need to stash arguments in locals to be able to call our method
on them (and still have them ready for the real call).
Added methods to ProfilerInfo for getting the sig and writing it back.