Add a server into the support assembly and create an initial "store".
I've added an ASP.NET Core web server. I suspect this may not be the long-term solution as it
just has too many dependencies (and I've had to put in an AssemblyResolve event hack to make
it work!).
Started organising the Support project.
Current state is that the server will spit out a list of active subscriptions in response to any
request.