Fix instrumentation bug.
If a method declared no locals and we needed to add some, we would fall over trying to get the
sig via a nil token.
Also fix the CHECK_SUCCESS macro so it actually prints the result and also prefixes with FAIL so
it's easy to search for in the log.