Implemented support for IConnectableObservable (probably v flaky)
Initial implementation relies on the support assembly taking a dependency on the Rx package, which
will no doubt lead to conflicts when running against a program that uses a different version of
Rx.

Possible remedies would be to do more AssemblyResolve stuff (we'd always want to resolve to the
version used by the program being monitored though), or to remove the dependency and support
IConnectableObservable (& IGroupedObservable, maybe others) by emitting dynamic wrapper types.