Add an event list screen to UI and make layout somewhat adjustable.
Add a new panel showing all events as we receive them. The events are in a data grid so they can
be copied to clipboard etc..

Provide a toolbar with buttons to pause and resume the event list update, and clear the event list.

Make the size of the three panels adjustable using grid splitters.

Some refactoring to put utility stuff in its own project.