Move to a multiple-tab arrangement in the UI (work in progress)
Define monitoring groups consisting of particular calls whose observables we're interested in.
Insert a tab control as the main view, with each tab corresponding to a monitoring group.
For now, add a new tab each time a call is chosen from the calls list.
Still to do: make it so that the content of each tab is actually different! Currently still showing
everything, as before.