Avoid instrumenting methods multiple times
Same method can be JITted several times (e.g. different instantiations of a generic method), and
we don't want to instrument it each time because we end up instrumenting our own instrumentation
code. So keep track of what we've instrumented and the code that we already produced and ensure
we set the same instrumented IL body & map each time.