Deal with args that are of interfaces derived from IObservable
Notably, IConnectableObservable was being wrapped when returned from a method, but also needs to
be recognised as an argument (notably to RefCount) so it can be linked to the result as an input.

I updated the code for delegate arguments (in case they return a derived interface) as well, but
I can't actually think of any method that this would be needed for.