Plug some gaps that were causing events to go missing.
Upon starting to monitor a chain, need to send existing subscription events for all the observables
down the chain, not just the top one.

An observable with no active subscriptions is not picked up when its IP starts being monitored, so
we need to check for this when a subscription is next created.

Remove the "don't wrap a wrapper" rule for observables. This was effectively "hiding" observables
from some instrumentation points, so attempting to monitor those points wouldn't do anything. It
would probably be more efficient to associate the existing wrapper with additional instrumentation
points but for now just let instances have multiple layers of wrapping.