Replace marble diagram with simple list of observable instances
Marble diagram for all observable instances from a given call site, and all inputs to those
instances, is too unwieldy to be the starting point, so replaced with a simple list of instances
(without inputs).

The intention is to reinstate the marble diagram but narrowed down to a selected instance and its
inputs.