Apply Pause/Go to all relevant screens.
Some issues: ideally when paused you should be able to switch to a different screen and see
the events that occurred up until the pause time, but because I've implemented at the view model
layer this isn't the case. Might need to redo at a lower level.

Though strangely, the observable list still shows stuff even when created while Paused, which
makes me suspect a bug in the Gate method's implementation.