Change marble diagram to use a TreeView
Lets us apply virtualization, which avoids the horrible pauses. However, it's still really hard to
use if there's a lot of data coming in quickly.

Also, probably need to change to a custom template as it looks a bit weird. Alternatively use a
ListBox/ListView and flatten the observables & subscriptions into one list in the view model,
since I'm not sure we really want the expand/collapse that TreeView provides.