Do pause/resume upstream of the model.
In other words, the model continues to reflect the state of play when it was paused, but screens
always show the latest from the model. This is better than continuing to update the model but
pausing the screens because (a) it's done in one place, and (b) the user can continue to navigate
between screens, and the screens they navigate to will show the state when the user hit pause,
rather than showing nothing until the user hits resume.