Replace ScrollViewers in each marble row with new SimpleViewport control
The new control just handles offsetting its child according to the specified offset values. This
avoids the annoying mousewheel behaviour of the ScrollViewer (it handles the events despite its
vertical scrolling being disabled, thus preventing the outer ScrollViewer from responding to them),
as well as being a lighter-weight control, which can't be bad.