Switch to specifying profiler by path
When launching program to profile, set COR_PROFILER_PATH_[32|64] and use the CLSID, rather than
relying on registry lookup.

Disable the automatic registration of the profiler when it's built (mainly so I can be sure it's
are working via the path mechanism).

Specifying Prefer32Bit=true parameter to TestProfilee now puts output in a different folder to
avoid confusion.