Remove sub from "active subs" on termination
Previously we only removed on disposal of the subscription, but this doesn't necessarily happen
even when the stream has terminated. We now remove when the stream terminates or when the
subscription is disposed (while still distinguishing those events).

This should ensure that the number of active subscriptions shown in the UI is a truer reflection
of what's going on.