Make data collector ask for instrumentation info
At the end of each test, ask for all instrumentation info and wait for it.

Was going to do this at the end of the session but that's too late because the process has been
shut down at that point.

Added concept of a "client event", which we use to record start and end of test cases, and also
as something to wait for to ensure we have everything we need from a test case before returning
from OnTestCaseEnd.

Need to add another mode to SendInstrumentationEventsRequest to get all events since the last
one sent (currently we're asking for everything-so-far after each test case, which is bad).