Start server as early as possible
Until the support assembly has been loaded (which should mean the server will be started in
short order), add the support assembly reference and module initializer to each module that
is loaded.

Additionally, in the module initializer, call a dummy method to force the class constructor
of the Instrument class to run (which starts the server).

In practice, we should normally end up doing this for just the executable assembly, since the class
initializer we add will force the support assembly to load before it gets to any method that
requires another module to be loaded.

Starting the server ASAP should mean that "test case start/end" events sent by the client appear
in the expected order relative to Rx-related events. It also means that we should have the problem
of the first test not doing any Rx stuff and thus not starting the server, leading to the test run
hanging as we wait for the response to the "test case end" event.