Work in progress
Configuration screen to show list of methods that the user has chosen, with the calls in them
(and indication as to which are being monitored).

Need to implement:
* List of methods in Workspace
* UI for selecting method(s) to add
* Interaction on the config screen (e.g. turn monitoring on/off for call, remove method from cfg)

Other changes here are general move towards "document screens" (tabs in UI). So config screen will
be such a document, and user will be able to open tabs to show e.g. event lists or diagrams for
particular calls/methods/observables.