Add dialog for adding method to config
Created a new method in IDialogService to show a dialog with general content, and implemented the
non-UI nuts and bolts in DialogService.

Hooked the UI part into the Shell (it should probably be a separate component but it'll do for
now). Dialog is shown as an overlay above the main content.

Implemented the add-method-to-config dialog using the above mechanism. This presents a list of
all the instrumented methods as a list, with a search box to filter down. The filtering is
currently somewhat crude.