Implement opening an event list document
i.e. show an event list for the current selection as a new tab.

Command is currently invoked only via the context menu (for a call in the monitoring configuration
or a selection of observables in the observables list).

Also a few fixes for bugs uncovered by monitoring configuration no longer being the only tab!