Bugfix: NRE in DynamicObservableAttacher
The attacher assumed that it would know which ObservableInfo it would be attaching the input to
by the time that the first subscription to the input happened. This is not the case, since the
method to which the delegate is passed could evaluate it and subscribe to the resulting input
observable before it returns (just after which AssociateWith is called).

So redo the logic in DynamicObservableAttacher so that if subscription happens first, we remember
that we need to call AddInput when AssociateWith is called.