Deal with cycles in input graph
In the EventListViewModel, the use of Expand to follow inputs would have caused problems if
any observable had itself as an input (possible via others). Additionally it wouldn't be very
efficient for input graphs that branch and merge a lot.

So added an ExpandDistinct method that's like Expand but filters out values that have been seen
already *before* invoking the recursive selector.