Redo connection screen
Remove process list and make the main content of the screen the launch parameters and controls to
set them.

Separate browsing for executable from launching so that user has the chance to set arguments and
options.

Make opening a data file a more obvious option.

Allow user to choose the "monitor from start" option that's used by test data collector (this also
sets the wait-for-connection option to ensure nothing is missed).